% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_mergenchat.R
\name{addin_mergenchat}
\alias{addin_mergenchat}
\title{Run Mergen Chat}
\usage{
addin_mergenchat(host = getOption("shiny.host", "127.0.0.1"))
}
\arguments{
\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
}
\value{
This function has no return value.
}
\description{
Run the Mergen Chat Shiny App as a background job and show it in the viewer pane
}
\examples{
# Call the function as an RStudio addin
if(interactive()){
addin_mergenchat()
}
}
