% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_mergenchat_insource.R
\name{addin_mergenchat_in_source}
\alias{addin_mergenchat_in_source}
\title{Mergen Chat in Source}
\usage{
addin_mergenchat_in_source()
}
\value{
This function has no return value.
}
\description{
Call this function as a Rstudio addin to ask mergen to improve spelling and
grammar of selected text.
}
