% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{check_api_key}
\alias{check_api_key}
\title{Check API key}
\usage{
check_api_key(api_key)
}
\arguments{
\item{api_key}{An API key.}
}
\description{
This function checks whether the API key provided as an argument is in the
correct format.
}
