% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir.linux}
\alias{choose.dir.linux}
\title{The linux version of the choose folder}
\usage{
choose.dir.linux(default = NA, caption = NA)
}
\arguments{
\item{default}{which folder to show initially}

\item{caption}{the caption on the selection dialog}
}
\description{
The linux version of the choose folder
}
\seealso{
\code{\link{choose.dir}}
}
