% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir.windows}
\alias{choose.dir.windows}
\title{The windows version of the choose folder}
\usage{
choose.dir.windows(default = NA, caption = NA, useNew = TRUE)
}
\arguments{
\item{default}{which folder to show initially}

\item{caption}{the caption on the selection dialog}

\item{useNew}{boolean, selects the type of dialog shown in windows}
}
\description{
The windows version of the choose folder
}
\seealso{
\code{\link{choose.dir}}
}
