% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{create_translator}
\alias{create_translator}
\title{Internationalization for the ChatGPT addin}
\usage{
create_translator(language = getOption("mergenstudio.language"))
}
\arguments{
\item{language}{The language to be found in the translation JSON file.}
}
\description{
The language can be set via \code{options("mergenstudio.language" = "<language>")}
(defaults to "en").
}
