% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_calls.R
\name{get_available_endpoints}
\alias{get_available_endpoints}
\title{List supported endpoints}
\usage{
get_available_endpoints()
}
\description{
Get a list of the endpoints supported by mergenstudio.
}
