% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_mergenchat_insource.R
\name{mergen_chat_completion}
\alias{mergen_chat_completion}
\title{Generate text completions using OpenAI's API for Chat}
\usage{
mergen_chat_completion(prompt, ai_api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{prompt}{The prompt for generating completions}

\item{ai_api_key}{The API key for accessing OpenAI's API. By default, the
function will try to use the \code{OPENAI_API_KEY} environment variable.}
}
\description{
Generate text completions using OpenAI's API for Chat
}
