% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_mergenchat_insource.R
\name{mergenchat_insource}
\alias{mergenchat_insource}
\title{Mergen Chat in Source}
\usage{
mergenchat_insource()
}
\value{
No return value. Replaces the selected text with new text provided by the request.
}
\description{
Giving more useful output in a source (i.e., *.R) file.
}
