% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_perform_request.R
\name{mergenstudio_request_perform.default}
\alias{mergenstudio_request_perform.default}
\title{#' @export
mergenstudio_request_perform.mergenstudio_request_palm <-
function(skeleton, ...) {
response <- create_completion_palm(prompt = skeleton$prompt)
structure(
list(
skeleton = skeleton,
response = response
),
class = "mergenstudio_response_palm"
)
}}
\usage{
\method{mergenstudio_request_perform}{default}(skeleton, ...)
}
\description{
#' @export
mergenstudio_request_perform.mergenstudio_request_anthropic <-
function(skeleton, ...) {
model   <- skeleton$model
prompt  <- skeleton$prompt
history <- skeleton$history
cli_inform(c("i" = "Using Anthropic API with {model} model"))
response <- create_completion_anthropic(prompt  = prompt,
history = history,
model   = model)
structure(
list(
skeleton = skeleton,
response = response
),
class = "mergenstudio_response_anthropic"
)
}
}
\details{
#' @export
mergenstudio_request_perform.mergenstudio_request_azure_openai <- function(skeleton, ...) {

messages <- c(
skeleton$history,
list(
list(role = "user", content = skeleton$prompt)
)
)

body <- list("messages"   = messages)

cat_print(body)

response <- create_completion_azure_openai(prompt = body)
structure(
list(
skeleton = skeleton,
response = response
),
class = "mergenstudio_response_azure_openai"
)
}
}
