% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_perform_request.R
\name{mergenstudio_request_perform}
\alias{mergenstudio_request_perform}
\title{Perform API Request}
\usage{
mergenstudio_request_perform(skeleton, ...)
}
\arguments{
\item{skeleton}{A \code{mergenstudio_request_skeleton} object}

\item{...}{Extra arguments (e.g., \code{stream_handler})}
}
\value{
A \code{mergenstudio_response_skeleton} object
}
\description{
This function provides a generic interface for calling different APIs
(e.g., OpenAI, HuggingFace, PALM (MakerSuite)). It dispatches the actual API
calls to the relevant method based on the \code{class} of the \code{skeleton} argument.
}
\examples{
\dontrun{
mergenstudio_request_perform(mergenstudio_skeleton)
}
}
