% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_process_response.R
\name{mergenstudio_response_process}
\alias{mergenstudio_response_process}
\title{Call API}
\usage{
mergenstudio_response_process(skeleton, ...)
}
\arguments{
\item{skeleton}{A \code{mergenstudio_response_skeleton} object}

\item{...}{Extra arguments, not currently used}
}
\value{
A \code{mergenstudio_request_skeleton} with updated history and prompt removed
}
\description{
This function provides a generic interface for calling different APIs
(e.g., OpenAI, PALM (MakerSuite)). It dispatches the actual API
calls to the relevant method based on the \code{class} of the \code{skeleton} argument.
}
\examples{
\dontrun{
mergenstudio_response_process(mergenstudio_skeleton)
}
}
