% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergenstudio-package.R
\docType{package}
\name{mergenstudio-package}
\alias{mergenstudio}
\alias{mergenstudio-package}
\title{mergenstudio: mergen Studio: An RStudio Addin wrapper for the mergen package}
\description{
An RStudio Addin wrapper for the mergen package.
}
\author{
\strong{Maintainer}: Artür Manukyan \email{artur-man@hotmail.com}

Authors:
\itemize{
  \item Jacqueline Jansen \email{jacqueline.jansen@uni-potsdam.de}
  \item Nour Alkhoury \email{nour.alkhoury@mdc-berlin.de}
  \item Altuna Akalin \email{altuna.akalin@mdc-berlin.de}
}

}
\keyword{internal}
