% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergenstudio-package.R
\docType{package}
\name{mergenstudio-package}
\alias{mergenstudio}
\alias{mergenstudio-package}
\title{mergenstudio: 'Mergen' Studio: An 'RStudio' Addin Wrapper for the 'Mergen' Package}
\description{
An 'RStudio' Addin wrapper for the 'mergen' package. This package employs artificial intelligence to convert data analysis questions into executable code, explanations, and algorithms. This package makes it easier to use Large Language Models in your development environment by providing a chat-like interface, while also allowing you to inspect and execute the returned code.
}
\author{
\strong{Maintainer}: Jacqueline Jansen \email{jacqueline-98@hotmail.com}

Authors:
\itemize{
  \item Artür Manukyan \email{artur-man@hotmail.com}
  \item Nour Alkhoury \email{nour.alkhoury@mdc-berlin.de}
  \item Altuna Akalin \email{altuna.akalin@mdc-berlin.de}
}

}
\keyword{internal}
